+++ backend/apps/cards/permissions.py
@@ -1,6 +1,6 @@
 from rest_framework.permissions import BasePermission
 
-from apps.cards.models import UserDeck, UserLeader, UserCard
+from apps.cards.models import UserCard, UserDeck, UserLeader
 
 
 class IsOwner(BasePermission):
