+++ backend/apps/cards/views.py
@@ -2,7 +2,7 @@ from rest_framework import mixins
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.viewsets import GenericViewSet
 
-from apps.cards.mixins import DeckBaseMixin, CardBaseMixin, LeaderBaseMixin
+from apps.cards.mixins import CardBaseMixin, DeckBaseMixin, LeaderBaseMixin
 from apps.cards.models import Card, Deck, Leader, UserCard, UserDeck, UserLeader
 from apps.cards.serializers import (
     CardSerializer,
